requirements.txt
asgiref==3.8.1
black==24.4.2
boto3==1.34.108
botocore==1.34.108
click==8.1.7
Django==5.0.6
flake8==7.0.0
jmespath==1.0.1
mccabe==0.7.0
mypy==1.10.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
pycodestyle==2.11.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
